Separate protos-to-generate from proto-deps
I think this is required to unblock https://github.com/reboot-dev/respect-api/issues/17,
where the generated code compiler might need a dependency list that's a
superset of the files it wants to generate code for. (In particular, the
set of output files != a 1:1 mapping to the set of dependencies: instead
it's a 1:1 mapping to the set of srcs).