Fix checks and add remote build cache (#25)
This supersedes https://github.com/reboot-dev/pyprotoc-plugin/pull/21 -
that PR was intended to correct incorrect workflow syntax, but revealed
that our tests couldn't run because the Python version on the GitHub
action runner machines is too low (3.8 vs our required 3.9).

This PR fixes the workflow syntax, installs Python 3.9 on the runner,
and while we're here also adds a remote build cache to make the build go
faster (originally ~7 minutes). The cache follows the now-standard model
where it is used read-only until a push to `main`, after which it is
updated.

Co-authored-by: Bulat Nizamutdinov <Bulat.Nizamutdinov@itechart-group.com>