Fix the CI by moving to `python3.10`
Before this PR our `pre-merge-checks.yml` would install `python3.9`.
Unfortunately that package has mysteriously disappeared from GitHub's
`ubuntu-latest`, which means our CI is broken.

This PR moves us to the `python3.10` package, which is what we use in
the rest of our repos anyway.

This is only a temporary fix; we're still beholden to GitHub's random
breakages as packages get removed, but it unbreaks our build and lets
us make progress.