Support `optional` fields
Before this PR, a proto file containing an `optional` field would cause
a pyprotoc plugin to fail, with an error along the lines of:

> tests/cpp/sample_service.proto: is a proto3 file that contains
> optional fields, but code generator protoc-gen-headers hasn't been
> updated to support optional fields in proto3. Please ask the owner of
> this code generator to support proto3 optional.--headers_out:oneof

This PR adds support for `optional` fields, by simply telling `protoc`
that we support them. Since `pyprotoc-plugin` doesn't itself generate
any code, actual support of `optional` fields is up to the developers
building pyprotoc plugins. Since `optional` fields have been a supported
feature of proto3 [for >2 years now](
https://github.com/protocolbuffers/protobuf/releases/tag/v3.15.0)
we believe that it is reasonable to assume that developers are aware of
the need to support them, just like any other type of field. Therefore
we believe it is reasonable to declare support for `optional` by
default (we also don't ask developers if they support `repeated`
fields, after all).

Fixes #29